EPMUII-7929 refactor: move IsTimeEqualWithHysteresis() and its test directly to equipment_test.go
